complete crud app

+++ notes-ui/notes-ui/Home.swift
@@ -15,6 +15,10 @@ struct Home: View {
     
     @State var showAlert = false
     @State var deleteItem: Note?
+    @State var updateNote = ""
+    @State var updateNoteId = ""
+    
+    @State var isEditMode: EditMode = .inactive
     
     var alert: Alert {
         Alert(title: Text("Delete"), message: Text("Are you sure you want to delete this note?"), primaryButton: .destructive(Text("Delete"), action: deleteNote), secondaryButton: .cancel())
@@ -23,25 +27,58 @@ struct Home: View {
     var body: some View {
         NavigationView {
             List(self.notes){ note in
-                Text(note.note)
-                    .padding()
-                    .onLongPressGesture {
-                        self.showAlert.toggle()
-                        deleteItem = note
+                if (self.isEditMode == .inactive){
+                    Text(note.note)
+                        .padding()
+                        .onLongPressGesture {
+                            self.showAlert.toggle()
+                            deleteItem = note
+                        }
+                }else{
+                    HStack{
+                        Image(systemName: "pencil.circle.fill")
+                            .foregroundColor(.yellow)
+                        
+                        Text(note.note)
+                            .padding()
                     }
+                    .onTapGesture {
+                        self.updateNote = note.note
+                        self.updateNoteId = note.id
+                        self.showAdd.toggle()
+                    }
+                }
             }
             .alert(isPresented: $showAlert, content: {
                 alert
             })
             .sheet(isPresented: $showAdd, onDismiss: fetchNotes
              , content: {
-                AddNoteView()
+                if (self.isEditMode == .inactive){
+                    AddNoteView()
+                } else {
+                    UpdateNoteView(text: $updateNote, noteId: $updateNoteId)
+                }
+                
+                
             })
             .onAppear(perform: {
                 fetchNotes()
             })
             .navigationTitle("Notes")
-            .navigationBarItems(trailing: Button(action: {
+            .navigationBarItems(leading: Button(action: {
+                if(self.isEditMode == .inactive){
+                    self.isEditMode = .active
+                } else{
+                    self.isEditMode = .inactive
+                }
+            }, label: {
+                if(self.isEditMode == .inactive){
+                    Text("Edit")
+                } else{
+                    Text("Done")
+                }
+            }) ,trailing: Button(action: {
                 self.showAdd.toggle()
             }, label: {
                 Text("Add")
@@ -66,6 +103,10 @@ struct Home: View {
             
         }
         task.resume()
+        
+        if (self.isEditMode == .active){
+            self.isEditMode = .inactive
+        }
     }
     
     func deleteNote(){
